# src/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple CRUD API for managing vehicle information.",
        "title": "Vehicle API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lekan Adetunmbi",
            "url": "https://pollaroid.net",
            "email": "me@pollaroid.net"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "api-vehicle.onrender.com",
    "basePath": "/v1",
    "paths": {
...
        "/vehicles": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get vehicles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "returns all vehicles",
                "responses": {
                    "200": {
                        "description": "all vehicles returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Vehicle"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "create vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "creates a new vehicle",
                "parameters": [
                    {
                        "description": "New Vehicle Details JSON",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateVehicleDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request body validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "409": {
                        "description": "another vehicle with the same model property exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
        "/vehicles/models": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get vehicle by model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "returns a vehicle by model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "vehcile search by model",
                        "name": "model",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "vehicle with the search model returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request query validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "no vehicle with the specified model found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        },
        "/vehicles/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get vehicle by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "returns a vehicle by its 16 caharcter uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "vehicle with the specified ID returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request param validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "vehicle with the specified ID not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "delete vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "deletes a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "vehicle deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.SuccessResponseDto"
                        }
                    },
                    "400": {
                        "description": "request param validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "update vehicle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicle"
                ],
                "summary": "updates a vehicle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID(UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vehicle Details JSON",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateVehicleDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "vehicle updated suucessfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.SuccessResponseDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Vehicle"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request body/param validation error or token not passed with request",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "401": {
                        "description": "invalid/expired token",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "404": {
                        "description": "vehicle with ID in request params not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "409": {
                        "description": "another vehicle with model in request body exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    },
                    "500": {
                        "description": "unexpected internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.FailedResponseDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.CreateVehicleDto": {
            "type": "object",
            "required": [
                "acceleration",
                "category",
                "cylinder",
                "horsePower",
                "imageUrl",
                "makeId",
                "model",
                "numberOfSeats",
                "price",
                "topSpeed",
                "torque",
                "transmission",
                "year"
            ],
            "properties": {
                "acceleration": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "cylinder": {
                    "type": "integer"
                },
                "horsePower": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "makeId": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "numberOfSeats": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "topSpeed": {
                    "type": "string"
                },
                "torque": {
                    "type": "string"
                },
                "transmission": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dtos.FailedResponseDto": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "errorType": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
        "dtos.SuccessResponseDto": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "statusText": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Make": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vehicle"
                    }
                }
            }
        },
...
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "acceleration": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "cylinder": {
                    "type": "integer"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "horsePower": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "make": {
                    "$ref": "#/definitions/models.Make"
                },
                "makeID": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "numberOfSeats": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "topSpeed": {
                    "type": "string"
                },
                "torque": {
                    "type": "string"
                },
                "transmission": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}